fix: fix generated UI tests and disable httpOnly on CSRF cookie
(cherry picked from commit cf485f5fdd0b19eed2d899f603c4479a3528e8d5)
